refactor: move properties
